Revert "build: 更新 C++ 标准和异常处理"
This reverts commit 3465e6ce9a5449f834564e2e467e35a2b9317326.

+++ src/core/components/sqlite_db.cpp
@@ -2,9 +2,8 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
-#include <cstddef> 
 
-#include "gdsqlite/sqlite/sqlite3.h"
+#include "godot-sqlist/src/sqlite/sqlite3.h"
 #include "sqlite_db.h"
 
 #include <godot_cpp/classes/file_access.hpp>
@@ -129,7 +128,7 @@ void SQLiteDB::Stmt::bind_double(int idx, double val) {
 	sqlite3_bind_double(stmt_, idx, val);
 }
 
-void SQLiteDB::Stmt::bind_blob(int idx, const void *val, size_t size) {
+void SQLiteDB::Stmt::bind_blob(int idx, const void *val, uint size) {
 	if (!stmt_) {
 		return;
 	}
